+++ CHANGELOG.md
@@ -100,3 +100,4 @@
 
 -   New login command with modal.
 -   Battle Eye / Anti Cheat command.
+-   Auto collect research points system job.